lesson 1, task №2: программа осуществялет математические операции на двумя переменными a и b, результаты выводятся на экран

+++ src/lecture1/MathOperations.java
@@ -12,10 +12,11 @@ public class MathOperations {
         System.out.println("Введите число b: ");
         double b = sc.nextDouble();
 
-        System.out.println("a + b = " + (double) ( a + b) );
-        System.out.println("a - b = " + (double) (a - b) );
-        System.out.println("a * b = " +  (double) a * b );
-        System.out.println("a / b = " +  (double) a / b );
+        System.out.println("a + b = " + (a + b) );
+        System.out.println("a - b = " + (a - b) );
+        System.out.println("a * b = " + (a * b) );
+        System.out.println("a / b = " + (a / b) );
 
     }
 }
+
